make tracked pagination default export
