+++ packages/pagination/pagination.stories.js
@@ -5,7 +5,7 @@ import { storiesOf } from "@storybook/react-native";
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { action } from "@storybook/addon-actions";
 import { withTrackingContext } from "@times-components/tracking";
-import Pagination, { withPageState, PaginationTracked } from "./pagination";
+import Pagination, { withPageState } from "./pagination";
 import { PreviousPageIcon, NextPageIcon } from "./pagination-icons";
 import LateralSpacingDecorator from "../../storybook/decorators/lateral-spacing";
 
@@ -62,12 +62,11 @@ storiesOf("Pagination", module)
       onPrev={action("last-page-prev-compact")}
     />
   ))
-  .add("With tracking", () => {
+  .add("Tracking", () => {
     const pageHandler = e => e.preventDefault();
-    const PaginationWithTrackingContext = withTrackingContext(
-      PaginationTracked,
-      { trackingObject: "Story" }
-    );
+    const PaginationWithTrackingContext = withTrackingContext(Pagination, {
+      trackingObject: "Story"
+    });
     return (
       <PaginationWithTrackingContext
         page={2}
